03-07-2019

+++ database/migrations/2019_06_11_032623_create_salaries_table.php
@@ -16,16 +16,16 @@ class CreateSalariesTable extends Migration
         Schema::create('salaries', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->bigInteger('employee_id')->unsigned();
-            $table->float('basic_pay')->nullable();
-            $table->float('house_rent')->nullable();
-            $table->float('dearness_allowance')->nullable();
-            $table->float('adhoc_2009')->nullable();
-            $table->float('adhoc_2010')->nullable();
-            $table->float('adhoc_2012')->nullable();
-            $table->float('adhoc_2013')->nullable();
-            $table->float('field_allowance')->nullable();
-            $table->float('other_allowance')->nullable();
-            $table->float('total');
+            $table->double('basic_pay')->nullable();
+            $table->double('house_rent')->nullable();
+            $table->double('dearness_allowance')->nullable();
+            $table->double('adhoc_2009')->nullable();
+            $table->double('adhoc_2010')->nullable();
+            $table->double('adhoc_2012')->nullable();
+            $table->double('adhoc_2013')->nullable();
+            $table->double('field_allowance')->nullable();
+            $table->double('other_allowance')->nullable();
+            $table->double('total');
             $table->bigInteger('promotion_id')->nullable();
             $table->text('salary_remarks')->nullable();
             $table->timestamps();
